Bug #9667 workaround: convert pthread_self to long before casting as unsigned int: this do not resolve the real problem (storing pointers as uint32 in 64bits pointers archs), but we preserve the API/ABI while making gcc more happy.
git-svn-id: svn+ssh://oberon/svn/OMiSCID-Dev@8846 c4bd1236-c4f7-0310-a2fe-a60683d9aa4a

+++ System/Thread.cpp
@@ -233,7 +233,7 @@ unsigned int Thread::GetThreadId()
 #ifdef WIN32
 	return (unsigned int)GetCurrentThreadId();
 #else
-	return (unsigned int)pthread_self();
+	return (unsigned int)((unsigned long)pthread_self());
 #endif
 }
